new updates scripts + readme

+++ api/autogame/tests/test_environment/tables/download_tables.py
@@ -18,21 +18,24 @@ def decode_item(item):
     else:
         return item
 
-# CLI prompt: python3 download_tables.py [dbHost] [dbName] [dbUser] [dbPass] [table_name]
+# CLI prompt: python3 download_tables.py [dbHost] [dbName] [dbUser] [dbPass] [table_name] [backup]
 if __name__ == "__main__":
     """
         This script downloads information from the db's tables so later the DB can be re-instantiated for testing.
-        Optional to pass a specific table_name to download
+        
+        Last argument its either [table_name] or [backup]:
+            * table_name: When running this script only by itself, its optional to pass a specific table_name to download
+            * backup: Indicates when this script is being called from the download_backup.py script
     """
 
     if len(sys.argv) >= 5:
 
-        output_directory = "./tables_files/"
+        output_directory = os.path.dirname(os.path.abspath(__file__)) + "/tables_files/"
 
         cnx = mysql.connector.connect(user=sys.argv[3], password=sys.argv[4], host=sys.argv[1], database=sys.argv[2])
         cursor = cnx.cursor(prepared=True)
 
-        if len(sys.argv) == 6:
+        if len(sys.argv) == 6 and sys.argv[5] != 'backup':
             # Delete file before downloading
             file_path = os.path.join(output_directory, sys.argv[5] + '.csv')
             os.remove(file_path)
@@ -54,7 +57,7 @@ if __name__ == "__main__":
         tables = cursor.fetchall()
 
         for table in tables:
-            if len(sys.argv) == 6:
+            if len(sys.argv) == 6 and sys.argv[5] != 'backup':
                 table_name = sys.argv[5]
             else:
                 table_name = table[0].decode()
@@ -77,19 +80,21 @@ if __name__ == "__main__":
                         decoded_row.append(decode_item(item))
                     csv_writer.writerow(decoded_row)
 
-            print(table_name + " - EXPORTED SUCCESSFULLY")
+            if (len(sys.argv) == 6 and sys.argv[5] != 'backup') or len(sys.argv) == 5:
+                print(table_name + " - EXPORTED SUCCESSFULLY")
 
-            if len(sys.argv) == 6:
+            if len(sys.argv) == 6 and sys.argv[5] != 'backup':
                 break
 
-        print("---------------------------------")
-        if len(sys.argv) == 6:
+        if len(sys.argv) == 6 and sys.argv[5] != 'backup':
             msg = "1 file exported"
 
         else:
             msg = str(len(tables)) + " file(s) exported."
 
-        print(msg)
+        if len(sys.argv) == 6 and sys.argv[5] != 'backup':
+            print("---------------------------------")
+            print(msg)
 
         cursor.close()
         cnx.close()
